requirements.txt
fastapi[standard]
black
python-dotenv
sqlmodel
pymysql
bcrypt
